% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_kath.R
\name{theme_kath}
\alias{theme_kath}
\title{theme_kath}
\usage{
theme_kath(..., base_size = 12)
}
\arguments{
\item{...}{allows additional theme settings}

\item{base_size}{base font size}
}
\value{
A ggplot2 theme
}
\description{
theme_kath
}
